1.13 Chunk Typings
I'm getting the typings for the untyped modules by typing them on my own with JSDoc, then I paste them into the TS Playground, and copy the output of the `.d.ts` file, after having set the output to CommonJS + ESNext.

Removing the main typings file, as it's not accurate to how the library is actually typed-out.